requirements.txt
flask>=2.3.0
gunicorn>=23.0.0
google-generativeai>=0.8.0
werkzeug>=2.3.0